Extract MinecraftLogObserver from InjectionConnection
