+++ Vanilla Injection/src/com/energyxxer/inject/v2/InjectionConnection.java
@@ -3,14 +3,12 @@ package com.energyxxer.inject.v2;
 import static com.energyxxer.inject.structures.StructureBlock.Mode.LOAD;
 import static com.energyxxer.inject.v2.CommandBlock.Type.CHAIN;
 import static com.energyxxer.inject.v2.CommandBlock.Type.REPEAT;
-import static com.google.common.base.Charsets.UTF_8;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.io.Files.createParentDirs;
 import static de.adrodoc55.minecraft.coordinate.Direction3.DOWN;
 import static java.nio.file.Files.isDirectory;
-import static java.nio.file.Files.isRegularFile;
 import static java.nio.file.StandardOpenOption.CREATE;
 import static java.nio.file.StandardOpenOption.READ;
 import static java.nio.file.StandardOpenOption.WRITE;
@@ -23,10 +21,8 @@ import java.nio.channels.OverlappingFileLockException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Collection;
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
@@ -43,11 +39,8 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.core.util.Integers;
 
-import com.energyxxer.inject.listeners.ChatEvent;
-import com.energyxxer.inject.listeners.LogEvent;
 import com.energyxxer.inject.listeners.SuccessEvent;
 import com.energyxxer.inject.structures.StructureBlock;
-import com.energyxxer.inject.utils.LogFileReader;
 import com.google.common.base.Charsets;
 import com.google.common.collect.Iterators;
 import com.google.common.primitives.Ints;
@@ -80,13 +73,8 @@ public class InjectionConnection implements AutoCloseable {
   private static final int CONNECTION_TIME_OUT = TIME_OUT_DELAY * TIME_OUT_CHECK_FREQUENCY;
 
   private final Logger logger;
+  private final MinecraftLogObserver logObserver;
 
-  /**
-   * The log file of the Minecraft installation. Note that when using multiple instances of
-   * Minecraft they will use the same log file and overwrite each others messages. For that reason
-   * an {@link InjectionConnection} only works properly when there is only one Minecraft instance.
-   */
-  private final Path logFile;
   /**
    * The directory of the Minecraft world.
    */
@@ -115,25 +103,10 @@ public class InjectionConnection implements AutoCloseable {
   private @Nullable FileChannel dataFileChannel;
 
   /**
-   * The {@link Thread} to use for the periodic actions ({@link #checkLog() log checking} and
-   * {@link #flush() flushing}). The {@link #executor} is {@code null} unless the connection
-   * {@link #isOpen()}.
+   * The {@link ScheduledExecutorService} to use for periodic {@link #flush() flushing}. The
+   * {@link #executor} is {@code null} unless the connection {@link #isOpen()}.
    */
   private @Nullable ScheduledExecutorService executor;
-  /**
-   * The {@link LogFileReader} to use for the periodic {@link #checkLog() log checking}. The
-   * {@link #reader} is {@code null} unless the connection {@link #isOpen()}.
-   */
-  private @Nullable LogFileReader reader;
-
-  /**
-   * The {@link ScheduledFuture} used to cancel periodic {@link #checkLog() log checking}. The
-   * {@link #logCheckFuture} is {@code null} unless the connection {@link #isOpen()}.
-   */
-  private @Nullable ScheduledFuture<?> logCheckFuture;
-  private long logCheckPeriod = 20;
-  private TimeUnit logCheckTimeUnit = MILLISECONDS;
-
   /**
    * The {@link ScheduledFuture} used to cancel periodic {@link #flush() flushing}. The
    * {@link #flushFuture} is {@code null} unless the connection {@link #isActive()}.
@@ -141,10 +114,6 @@ public class InjectionConnection implements AutoCloseable {
   private @Nullable ScheduledFuture<?> flushFuture;
   private long flushPeriod = 20;
   private TimeUnit flushTimeUnit = MILLISECONDS;
-
-  private final List<Consumer<? super LogEvent>> logListeners = new CopyOnWriteArrayList<>();
-  private final List<Consumer<? super ChatEvent>> chatListeners = new CopyOnWriteArrayList<>();
-  private final Collection<SuccessListenerEntry> successListeners = new ConcurrentLinkedQueue<>();
 
   /**
    * The ID of the next {@link Structure} that will be generated by {@link #flush()}. This ID is
@@ -165,8 +134,8 @@ public class InjectionConnection implements AutoCloseable {
 
   /**
    * This flag indicates whether or not the next flush operation should generate commands that if
-   * successful are written to the log file. This is usually set to {@code true} when listening for
-   * the output of some commands.
+   * successful are written to the {@link MinecraftLogObserver#logFile log file}. This is usually
+   * set to {@code true} when listening for the output of some commands.
    * <p>
    * <b>Implementation Notes</b><br>
    * This field is not {@code volatile} because it is only read while holding the exclusive write
@@ -193,10 +162,9 @@ public class InjectionConnection implements AutoCloseable {
   private final Collection<Command> commandBuffer = new ConcurrentLinkedQueue<>();
 
   /**
-   * Creates and {@link #open() opens} a new {@link InjectionConnection} with the specified
-   * parameters.
+   * Create and {@link #open()} a new {@link InjectionConnection} with the specified parameters.
    *
-   * @param logFile the {@link #logFile} of the Minecraft installation
+   * @param logFile the {@link MinecraftLogObserver#logFile log file} of the Minecraft installation
    * @param worldDir the {@link #worldDir} of the Minecraft world
    * @param identifier the unique {@link #identifier} for {@code this} connection in the Minecraft
    *        world
@@ -206,12 +174,11 @@ public class InjectionConnection implements AutoCloseable {
    */
   public InjectionConnection(Path logFile, Path worldDir, String identifier)
       throws IOException, InterruptedException {
-    this.logFile = checkNotNull(logFile, "logFile == null!");
     this.worldDir = checkNotNull(worldDir, "worldDir == null!");
-    checkArgument(isRegularFile(logFile), "%s is not a regular file!", logFile);
     checkArgument(isDirectory(worldDir), "%s is not a directory!", worldDir);
     this.identifier = checkNotNull(identifier, "identifier == null!");
     logger = LogManager.getLogger(toString());
+    logObserver = new MinecraftLogObserver(logFile);
     structureDir = worldDir.resolve("structures");
     injectionDir = structureDir.resolve("inject").resolve(identifier);
     dataFile = injectionDir.resolve("data.txt");
@@ -228,11 +195,19 @@ public class InjectionConnection implements AutoCloseable {
   }
 
   /**
+   * @return the value of {@link #logObserver}
+   */
+  public MinecraftLogObserver getLogObserver() {
+    return logObserver;
+  }
+
+  /**
    * {@link #isOpen() Open} {@code this} connection and wait for a response from Minecraft. Once the
    * response is received, {@link #isActive() activate} {@code this} connection.
    *
-   * @throws IOException if an I/O error occurs while opening the {@link #logFile}, locking the
-   *         {@link #counterFile} or {@link #flush() flushing}
+   * @throws IOException if an I/O error occurs while opening the
+   *         {@link MinecraftLogObserver#logFile log file}, locking the {@link #counterFile} or
+   *         {@link #flush() flushing}
    * @throws InterruptedException if the current thread is interrupted while waiting for Minecraft's
    *         response
    */
@@ -240,12 +215,11 @@ public class InjectionConnection implements AutoCloseable {
     checkState(!isOpen(), "This connection is already established!");
     logger.info("Establishing connection");
     lockDataFile();
+    logObserver.open();
+    executor = Executors.newSingleThreadScheduledExecutor();
     int structureId = loadStructureId();
     this.structureId.set(structureId);
     logger.info("Using structure '{}'", getStructureName(structureId));
-    executor = Executors.newSingleThreadScheduledExecutor();
-    reader = new LogFileReader(logFile);
-    schedulePeriodicLogCheck();
     Semaphore semaphore = new Semaphore(0);
     injectCommand(SUCCESSFUL_COMMAND, e -> {
       confirmStructure(structureId);
@@ -284,11 +258,6 @@ public class InjectionConnection implements AutoCloseable {
     throw new IllegalStateException("A " + this + " is already open", ex);
   }
 
-  private void closeDataFileChannel() throws IOException {
-    dataFileChannel.close();
-    dataFileChannel = null;
-  }
-
   /**
    * Load the value for {@link #structureId} from {@link #dataFile} .
    *
@@ -307,21 +276,25 @@ public class InjectionConnection implements AutoCloseable {
   public void close() throws IOException {
     if (isOpen()) {
       logger.info("Closing connection");
+      logObserver.close();
       if (isActive()) {
         cancelPeriodicFlush();
       }
-      cancelPeriodicLogCheck();
       executor.shutdown();
       executor = null;
-      reader = null;
       lastConfirmedStructureId = -1;
       flush();
       closeDataFileChannel();
     }
   }
 
+  private void closeDataFileChannel() throws IOException {
+    dataFileChannel.close();
+    dataFileChannel = null;
+  }
+
   /**
-   * Return whether {@code this} connection periodically {@link #checkLog() checks the log file}.
+   * Return whether {@code this} connection {@link #isActive()} or is ready to be activated.
    * <p>
    * A connection is opened after construction or by calling {@link #open()}.<br>
    * To {@link #isClosed() close} {@code this} connection call {@link #close()}.
@@ -329,7 +302,7 @@ public class InjectionConnection implements AutoCloseable {
    * @return whether {@code this} connection is open
    */
   public boolean isOpen() {
-    return executor != null && reader != null && logCheckFuture != null && dataFileChannel != null;
+    return executor != null && dataFileChannel != null;
   }
 
   /**
@@ -346,7 +319,7 @@ public class InjectionConnection implements AutoCloseable {
 
   /**
    * Pause {@code this} connection by canceling periodic {@link #flush() flushing}. Pausing does not
-   * stop periodic {@link #checkLog() log checking}.<br>
+   * stop periodic {@link MinecraftLogObserver#checkLog() log checking}.<br>
    * After pausing {@link #isOpen()} will return {@code true} and {@link #isActive()} will return
    * {@code false}.
    *
@@ -415,11 +388,6 @@ public class InjectionConnection implements AutoCloseable {
     return isOpen() && !isActive();
   }
 
-  private void schedulePeriodicLogCheck() {
-    logCheckFuture =
-        executor.scheduleAtFixedRate(this::checkLog, 0, logCheckPeriod, logCheckTimeUnit);
-  }
-
   private void schedulePeriodicFlush() {
     flushFuture = executor.scheduleAtFixedRate(() -> {
       try {
@@ -440,43 +408,9 @@ public class InjectionConnection implements AutoCloseable {
     }, 0, flushPeriod, flushTimeUnit);
   }
 
-  private void cancelPeriodicLogCheck() {
-    logCheckFuture.cancel(false);
-    logCheckFuture = null;
-  }
-
   private void cancelPeriodicFlush() {
     flushFuture.cancel(false);
     flushFuture = null;
-  }
-
-  /**
-   * @return the value of {@link #logCheckPeriod}
-   */
-  public long getLogCheckPeriod() {
-    return logCheckPeriod;
-  }
-
-  /**
-   * @return the value of {@link #logCheckTimeUnit}
-   */
-  public TimeUnit getLogCheckTimeUnit() {
-    return logCheckTimeUnit;
-  }
-
-  /**
-   * Set the frequency at which the {@link #logFile} is checked for changes.
-   *
-   * @param logCheckPeriod the new value for {@link #logCheckPeriod}
-   * @param logCheckTimeUnit the new value for {@link #logCheckTimeUnit}
-   */
-  public void setLogCheckFrequency(long logCheckPeriod, TimeUnit logCheckTimeUnit) {
-    this.logCheckPeriod = logCheckPeriod;
-    this.logCheckTimeUnit = checkNotNull(logCheckTimeUnit, "logCheckTimeUnit == null!");
-    if (logCheckFuture != null) { // isOpen()
-      cancelPeriodicLogCheck();
-      schedulePeriodicLogCheck();
-    }
   }
 
   /**
@@ -502,99 +436,10 @@ public class InjectionConnection implements AutoCloseable {
   public void setFlushFrequency(long flushPeriod, TimeUnit flushTimeUnit) {
     this.flushPeriod = flushPeriod;
     this.flushTimeUnit = checkNotNull(flushTimeUnit, "flushTimeUnit == null!");
-    if (flushFuture != null) { // isActive()
+    if (isActive()) {
       cancelPeriodicFlush();
       schedulePeriodicFlush();
     }
-  }
-
-  /**
-   * Check for new changes to the {@link #logFile} since the last check or alternatively since
-   * {@link #open() opening} {@code this} connection and dispatch new events.
-   * <p>
-   * If {@code this} connection {@link #isActive()} a log check will be performed periodically
-   * according to {@link #logCheckPeriod} and {@link #logCheckTimeUnit}.
-   */
-  private void checkLog() {
-    reader.readAddedLines(UTF_8, this::handleLogLine);
-  }
-
-  private void handleLogLine(String line) {
-    dispatchLogEvent(new LogEvent(line));
-    ChatEvent ce = ChatEvent.createFromLogLine(line);
-    if (ce != null) {
-      dispatchChatEvent(ce);
-    }
-    SuccessEvent se = SuccessEvent.createFromLogLine(line);
-    if (se != null) {
-      dispatchSuccessEvent(se);
-    }
-  }
-
-  private void dispatchLogEvent(LogEvent event) {
-    for (Consumer<? super LogEvent> listener : logListeners) {
-      listener.accept(event);
-    }
-  }
-
-  /**
-   * Add the specified listener to be notified for every new line in the {@link #logFile}.
-   *
-   * @param listener
-   */
-  public void addLogListener(Consumer<? super LogEvent> listener) {
-    logListeners.add(listener);
-  }
-
-  /**
-   * Remove the specified log listener.
-   *
-   * @param listener
-   * @return {@code true} if the specified listener was previously registered
-   */
-  public boolean removeLogListener(Consumer<? super LogEvent> listener) {
-    return logListeners.remove(listener);
-  }
-
-  private void dispatchChatEvent(ChatEvent event) {
-    for (Consumer<? super ChatEvent> listener : chatListeners) {
-      listener.accept(event);
-    }
-  }
-
-  /**
-   * Add the specified listener to be notified whenever a chat event is detected in the
-   * {@link #logFile}.
-   *
-   * @param listener
-   */
-  public void addChatListener(Consumer<? super ChatEvent> listener) {
-    chatListeners.add(listener);
-  }
-
-  /**
-   * Remove the specified chat listener.
-   *
-   * @param listener
-   * @return {@code true} if the specified listener was previously registered
-   */
-  public boolean removeChatListener(Consumer<? super ChatEvent> listener) {
-    return chatListeners.remove(listener);
-  }
-
-  private void dispatchSuccessEvent(SuccessEvent event) {
-    for (SuccessListenerEntry entry : successListeners) {
-      if (event.getInvoker().equals(entry.getInvoker())) {
-        if (!entry.isRepeat()) {
-          successListeners.remove(entry);
-        }
-        entry.getListener().accept(event);
-      }
-    }
-  }
-
-  private void addSuccessListener(String invoker, boolean repeat, Consumer<SuccessEvent> listener) {
-    successListeners.add(new SuccessListenerEntry(invoker, repeat, listener));
   }
 
   /**
@@ -738,24 +583,27 @@ public class InjectionConnection implements AutoCloseable {
     dataFileChannel.truncate(dataFileChannel.position());
   }
 
-  public void injectCommand(String command) {
+  public void injectCommand(String command) throws IllegalStateException {
     injectCommand(new Command(command));
   }
 
-  public void injectCommand(Command command) {
+  public void injectCommand(Command command) throws IllegalStateException {
+    checkOpen();
     commandBuffer.add(command);
   }
 
-  public void injectCommand(String command, Consumer<SuccessEvent> listener) {
+  public void injectCommand(String command, Consumer<SuccessEvent> listener)
+      throws IllegalStateException {
     String name = UUID.randomUUID().toString();
     injectCommand(new Command(name, command), listener);
   }
 
-  public void injectCommand(Command command, Consumer<SuccessEvent> listener) {
+  public void injectCommand(Command command, Consumer<SuccessEvent> listener)
+      throws IllegalStateException {
     logAdminCommandsLock.readLock().lock();
     try {
       injectCommand(command);
-      addSuccessListener(command.getName(), false, listener);
+      logObserver.addSuccessListener(command.getName(), false, listener);
       logAdminCommands = true;
     } finally {
       logAdminCommandsLock.readLock().unlock();
